Debian packaging: install /etc/default/dcs-client
Pull the configuration items that the user must set (that is, locations
of various keys and certificates) into environment variables set in
/etc/default/dcs-client. This file could still benefit from some
documenting comments.

Also set the variable ENABLED="0" by default, so that on installation the
dcs-client service is disabled, rather than failing to start up due to
incomplete config.