Debian packaging: upstart fixes
Don't try to create /var/log/dcs-client in the pre-start script, because
upstart doesn't allow us to run the pre-start script as root and the
start script as dcs-client. This puts the responsibility of creating and
chmoding /var/log/dcs-client onto the postinst maintainer script.

Copy the variables in /etc/default/dcs-client into the environment
before exec'ing Java, so that dcs-client can see them.