Replace Dockerfile with gradle-docker task
Use gradle-docker to generate the Dockerfile at build time, rather than
using a static Dockerfile with a VERSION argument. This eliminates the
ARG and ENV parts of the Dockerfile.

Use `openjdk:8-jre-alpine` as the base image instead of `alpine:3.5`.
This eliminates the `apk` commands from the Dockerfile and speeds up the
build.